Isolate the whole application state into the root AppImpl
Isolate the whole application state in the AppImpl, with support for
instantiation of multiple simultaneous root AppImpls each with their
own trace loaded and rendered in their own UI.

Multiple registration of plugins in separate plugin managers needs to
tolerate attempts to register the same feature flags again, so that
is accounted for. All app instances share the same feature flags, which
makes sense as they are maintained in session storage.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>