ui: isolate colors, fonts, etc into overridable CSS variables
Isolate colors, fonts, and other definitions into CSS variables that can
be overridden by a host application to apply its themes.

Also replace the in-line SVG icon in the tree widget (which had a stroke
colour that could not be set from CSS) with the material icon that is a
chevron shape.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>